voice_agent/requirements.txt
fastapi==0.103.2
uvicorn==0.24.0
pydantic==2.5.3
pydantic-settings==2.0.3
sqlalchemy==2.0.23
fastapi-users[sqlalchemy]==10.4.2
alembic==1.13.1
celery==5.3.6
redis==5.0.3
openai-whisper
pydub==0.25.1
portkey==0.1.0

# Install PyTorch and TorchAudio with CUDA 11.8 support
https://download.pytorch.org/whl/cu118/torch-2.3.0%2Bcu118-cp310-cp310-linux_x86_64.whl
https://download.pytorch.org/whl/cu118/torchaudio-2.3.0%2Bcu118-cp310-cp310-linux_x86_64.whl

git+https://github.com/SWivid/F5-TTS.git
python-jose[cryptography]==3.3.0
passlib==1.7.4
httpx==0.25.0
pre-commit==3.5.0
ruff==0.1.6
black==23.10.0